Removed duplicated handling of unset commands
Removed the handling from the cli code and instead made it catch
exceptions raised by the handling code from the library functions.

This does mean that if e.g. the blender command is not set, mcrender
will first complete the mineways step before notifying the user of the
problem. Since an unset command should rarely occur, this is an
acceptable loss.